requirements-cpu.txt
biopython==1.81
fair-esm==2.0.0
filelock==3.13.1
fsspec==2023.10.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
sympy==1.12
torch==2.1.0
typing_extensions==4.8.0
